Skills/Stock/HongKong_Stock/ma-bullish-strategy/requirements.txt
akshare>=1.12.0
pandas>=1.5.0
numpy>=1.23.0
pyyaml>=6.0
tushare>=1.3.0
